fixed existing test failures
